Removed old config files.
