+++ DEA/appconfig.json
@@ -20,9 +20,11 @@
   "FoldersToExclude": [
     {
       "MainEmailFolders": [
-        "Level 2"
+        "Level 1",
+        "Level 3"
       ],
       "SubEmailFolders": [
+        "Sub 1"
       ]
     }    
   ]
