--- DEA/dea.conf
@@ -1,6 +1,0 @@
-﻿DateFilter = false
-MaxLoadEmails = 4
-UserAccounts = sandaruwan@efakturamottak.no
-ImportFolderLetter = D:
-ImportFolderPath = Import
-AllowedExtentions = .jpg, .jpeg, .tiff, .png, .pdf
